+++ src/main/java/br/ufsm/csi/dao/PlayerDAO.java
@@ -37,6 +37,33 @@ public class PlayerDAO {
         return users;
     }
 
+    public Player getPlayerById(int id){
+        Player userFound = new Player();
+
+        try(Connection connection = new ConectDB().getConexao()){
+            this.sql = "SELECT * FROM player";
+
+            this.statement = connection.createStatement();
+            this.resultSet = this.statement.executeQuery(sql);
+
+            while (this.resultSet.next()) {
+                if (this.resultSet.getInt("userid") == id) {
+                    Player user = new Player();
+                    user.setUserId(this.resultSet.getInt("userid"));
+                    user.setName(this.resultSet.getString("name"));
+                    user.setUsername(this.resultSet.getString("username"));
+
+                    userFound = user;
+                }
+            }
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
+        return userFound;
+    }
+
     public String insert(Player user) {
         try(Connection connection = new ConectDB().getConexao()) {
 
@@ -59,11 +86,13 @@ public class PlayerDAO {
                 user.setUserId(this.resultSet.getInt(1));
                 connection.commit();
                 this.status = "OK";
+                System.out.println("* - Player added with success - *");
             }
 
         } catch (SQLException e) {
             e.printStackTrace();
             this.status = "ERROR";
+            System.out.println("* - Could not add Player - *");
         }
 
         return this.status;
@@ -89,10 +118,10 @@ public class PlayerDAO {
             if (ok > 0) {
                 connection.commit();
                 this.status = "OK";
-                System.out.println("Atualizado com sucesso! !");
+                System.out.println("* - Player was updated with success - *");
             } else {
                 this.status = "ERROR";
-                System.out.println("Ocorreu um erro ao atualizar! !");
+                System.out.println("x - Player update failed - x");
             }
 
         } catch (SQLException e) {
@@ -108,25 +137,40 @@ public class PlayerDAO {
             //BEGIN
             connection.setAutoCommit(false);
 
-            //delete from id
-
-            this.sql = "delete from player " +
-                    " where userid = ?";
+            this.sql = "DELETE FROM usercharacter " +
+                    "WHERE usercharacter.userid = ?";
 
             this.preparedStatement = connection.prepareStatement(this.sql);
             this.preparedStatement.setInt(1, user.getUserId());
 
-            int ok2 = this.preparedStatement.executeUpdate();
-
-            if (ok2 > 0) {
+            int deletedFromCharacter = this.preparedStatement.executeUpdate();
+            if (deletedFromCharacter > 0) {
                 connection.commit();
                 this.status = "OK";
-                System.out.println("Usuario deletado com sucesso! !");
+                System.out.println("* - [Player] was removed from [Character] - *");
             } else {
+                System.out.println(" x - [Player] could not be removed from [Character] - x");
                 this.status = "ERROR";
-                System.out.println("Usuario nao existe.");
             }
 
+            if(this.status.equals("OK")) {
+                this.sql = "delete from player " +
+                        " where userid = ?";
+
+                this.preparedStatement = connection.prepareStatement(this.sql);
+                this.preparedStatement.setInt(1, user.getUserId());
+
+                int ok2 = this.preparedStatement.executeUpdate();
+
+                if (ok2 > 0) {
+                    connection.commit();
+                    this.status = "OK";
+                    System.out.println("* - Player was deleted with success - *");
+                } else {
+                    this.status = "ERROR";
+                    System.out.println("x - Delete player failed - x");
+                }
+            }
 
         } catch (SQLException e) {
             e.printStackTrace();
